+++ cayleypy/puzzles/__init__.py
@@ -1,2 +1,3 @@
 from .cube import rubik_cube
 from .globe import globe_puzzle
+from .hungarian_rings import hungarian_rings_permutations, hungarian_rings_generators
