+++ cayleypy/puzzles/hungarian_rings.py
